+++ packages/pangea-sdk/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Vault `/export` support.
+- AuthN user password expiration support.
 
 ## [3.9.0] - 2024-06-07